backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
SQLAlchemy==2.0.30
alembic==1.13.1
psycopg[binary,pool]==3.1.19
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.7
prometheus-client==0.20.0
sqlglot==25.13.0
httpx==0.27.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
PyJWT[crypto]==2.9.0
